add ui for tracking phases
