+++ crates/thetawave_interface/Cargo.toml
@@ -16,3 +16,5 @@ bevy_reflect = { workspace = true }
 leafwing-input-manager = { workspace = true }
 bevy_input = { workspace = true }
 ron = { workspace = true }
+bevy_time = { workspace = true }
+bevy_math = { workspace = true }
