+++ src/collision/contact.rs
@@ -1,12 +1,12 @@
 use crate::{
     arena::ArenaBarrierComponent,
-    player::PlayerComponent,
     spawnable::{MobComponent, MobSegmentComponent, ProjectileComponent},
 };
 use bevy::prelude::*;
 use bevy_rapier2d::prelude::*;
 use thetawave_interface::{
     audio::{CollisionSoundType, PlaySoundEffectEvent, SoundEffectType},
+    player::PlayerComponent,
     spawnable::{Faction, MobSegmentType, MobType, ProjectileType},
 };
