--- src/player/components/mod.rs
@@ -1,5 +1,0 @@
-//! Components for managing players
-
-mod player;
-
-pub use self::player::*;
