+++ src/player/mod.rs
@@ -8,7 +8,6 @@ use thetawave_interface::{
     states::{AppStates, GameStates},
 };
 
-mod components;
 mod resources;
 mod spawn;
 mod systems;
@@ -16,8 +15,7 @@ mod systems;
 use crate::{GameEnterSet, GameUpdateSet};
 
 pub use self::{
-    components::PlayerComponent,
-    resources::{Character, CharactersResource},
+    resources::CharactersResource,
     spawn::spawn_players_system,
     systems::{
         player_ability_system, player_death_system, player_fire_weapon_system,
