+++ src/spawnable/mob/behavior.rs
@@ -4,7 +4,8 @@ use bevy_rapier2d::prelude::*;
 use serde::Deserialize;
 use thetawave_interface::{
     audio::{PlaySoundEffectEvent, SoundEffectType},
-    health::DamageDealtEvent,
+    health::{DamageDealtEvent, HealthComponent},
+    player::PlayerComponent,
     spawnable::{EffectType, MobType, ProjectileType},
 };
 
@@ -12,11 +13,7 @@ use crate::{
     collision::SortedCollisionEvent,
     game::GameParametersResource,
     loot::LootDropsResource,
-    misc::HealthComponent,
-    spawnable::{
-        InitialMotion, PlayerComponent, SpawnConsumableEvent, SpawnEffectEvent,
-        SpawnProjectileEvent,
-    },
+    spawnable::{InitialMotion, SpawnConsumableEvent, SpawnEffectEvent, SpawnProjectileEvent},
 };
 
 use super::{MobComponent, SpawnMobEvent, SpawnPosition};
