+++ src/spawnable/mob/mob_segment/mod.rs
@@ -4,6 +4,7 @@ use serde::Deserialize;
 use std::collections::{hash_map::Entry, HashMap};
 use thetawave_interface::{
     audio::CollisionSoundType,
+    health::HealthComponent,
     objective::DefenseInteraction,
     spawnable::{MobSegmentType, SpawnableType},
     states::GameCleanup,
@@ -14,7 +15,6 @@ use crate::{
     assets::MobAssets,
     game::GameParametersResource,
     loot::ConsumableDropListType,
-    misc::HealthComponent,
     spawnable::SpawnableComponent,
     HORIZONTAL_BARRIER_COL_GROUP_MEMBERSHIP, SPAWNABLE_COL_GROUP_MEMBERSHIP,
 };
