+++ src/spawnable/mob/mod.rs
@@ -12,7 +12,6 @@ use crate::{
     assets::MobAssets,
     game::GameParametersResource,
     loot::ConsumableDropListType,
-    misc::HealthComponent,
     spawnable::{InitialMotion, SpawnableBehavior, SpawnableComponent},
     HORIZONTAL_BARRIER_COL_GROUP_MEMBERSHIP, SPAWNABLE_COL_GROUP_MEMBERSHIP,
 };
@@ -24,6 +23,7 @@ pub use self::{behavior::*, mob_segment::*};
 use super::behavior_sequence::MobBehaviorSequenceType;
 use thetawave_interface::{
     audio::CollisionSoundType,
+    health::HealthComponent,
     objective::DefenseInteraction,
     spawnable::{MobSegmentType, MobType, ProjectileType},
     states::GameCleanup,
